+++ tests/BoltOn.Tests/Mediator/GetStudentHandler.cs
@@ -1,8 +1,9 @@
 ﻿using System.Linq;
 using BoltOn.Data;
 using BoltOn.Mediator.Pipeline;
+using BoltOn.Tests.Other;
 
-namespace BoltOn.Tests.Other
+namespace BoltOn.Tests.Mediator
 {
 	public class GetStudentRequest : IQuery<Student>
 	{
