+++ tests/BoltOn.Tests/Mediator/MediatorTestHelper.cs
@@ -1,14 +1,12 @@
 ﻿using System.Collections.Generic;
 
-namespace BoltOn.Tests.Other
+namespace BoltOn.Tests.Mediator
 {
 	public class MediatorTestHelper
     {
         public static List<string> LoggerStatements { get; set; } = new List<string>();
         public static bool IsClearInterceptors { get; set; }
         public static bool IsCustomizeIsolationLevel { get; set; }
-		public static bool IsSeedData { get; set; } = true;
-		public static bool IsSqlServer { get; set; }
 		public static bool IsRemoveStopwatchInterceptor { get; set; }
 	}
 }
