+++ tests/BoltOn.Tests/Other/IntegrationTestHelper.cs
@@ -4,5 +4,7 @@ namespace BoltOn.Tests.Other
 	{
 		public static bool IsSqlRunning { get; set; } = false;
 		public static bool IsRabbitMqRunning { get; set; } = false;
+		public static bool IsSqlServer { get; set; }
+		public static bool IsSeedData { get; set; } = true;
 	}
 }
